+++ natureoverhaul/TreeData.java
@@ -68,11 +68,11 @@ public class TreeData {
 
     @Override
     public boolean equals(Object obj){
-        if(obj==null){
-            return false;
-        }
         if(obj==this){
             return true;
+        }
+        if(obj==null){
+            return false;
         }
         if(obj instanceof TreeData){
             for(Component component:Component.values()){