+++ natureoverhaul/behaviors/BehaviorStarving.java
@@ -20,7 +20,7 @@ public abstract class BehaviorStarving extends Behavior {
 
 	@Override
 	public boolean hasDied(World world, int i, int j, int k, Block id) {
-		return NatureOverhaul.useStarvingSystem && hasStarved(world, i, j, k, id);
+		return NatureOverhaul.INSTANCE.useStarvingSystem && hasStarved(world, i, j, k, id);
 	}
 
 	/**